+++ meta_package_manager/tests/test_definitions.py
@@ -60,11 +60,11 @@ class TestDefinition(unittest.TestCase):
     def test_cli_type(self):
         """ Check that all methods returning a CLI is a list. """
         for manager in pool().values():
-            self.assertIsInstance(manager.update_cli('dummy_package_id'), list)
+            self.assertIsInstance(manager.upgrade_cli('dummy_package_id'), list)
 
-            # Update-all CLI are allowed to raise a particular error.
+            # Upgrade-all CLI are allowed to raise a particular error.
             try:
-                result = manager.update_all_cli()
+                result = manager.upgrade_all_cli()
             except Exception as excpt:
                 self.assertIsInstance(excpt, NotImplementedError)
             else: